refactor(blocks-engine): move the breakpoint-set helpers into the engine
`authoredBreakpoints` and `inCascadeOrder` answer questions a stored
`BreakpointSet` does not: which rows an author actually defined, and the order
the cascade applies them in. Both were in the editor because the editor asked
first, and both are pure functions over types the engine already owns.

They now have a second asker outside that package — a preview surface deriving
device presets from the same record — and a second implementation would agree
about what a breakpoint MEANS while disagreeing about which rows exist and in
what order. That is the quieter half of the same defect: a stored set may carry
a reserved `base` row the compiler prepends regardless, so a consumer that skips
the strip shows an author a "Base" device that is not one, in stored rather than
cascade order.

Placed beside `breakpoint-axes.ts` for the reason that file exists: facts about
the document model, readable without reading the emitter that consumes them.

The editor re-exports both rather than keeping copies, which is the convention
its own module header already states for every value the compiler decides. Every
existing import is unchanged, including the plugin's.

`inCascadeOrder` now takes a readonly array. It already copied before sorting —
`sort` mutates, and sorting a caller's stored record reorders it for every other
reader — and the parameter type now says so rather than leaving it to the body.

Their tests move with them. The engine suite gains four and the builder suite
loses the same four, which is the deliberate relocation the count rule exempts.
