+++ packages/builder/src/breakpoints.test.ts
@@ -100,38 +100,6 @@ const BREAKPOINTS_FOR_PREVIEW = {
   container: [],
 } as unknown as BreakpointSet;
 
-describe("the authored set", () => {
-  it("strips a stored base row from both axes", () => {
-    /*
-     * A stored set CAN carry a `base` row and the plugin's README documents a
-     * host config that does, while the compiler claims that id before reading
-     * any stored definition. Three surfaces ask "what has this site defined" —
-     * the dialog's draft, the trigger's count, and the host deciding whether
-     * config defaults exist — and each one that answered differently produced
-     * its own defect.
-     */
-    const stripped = authoredBreakpoints({
-      viewport: [
-        { id: "base", label: "Base" },
-        { id: "tablet", label: "Tablet", maxWidth: 991 },
-      ],
-      container: [{ id: "base", label: "Base" }],
-    } as unknown as BreakpointSet);
-
-    expect(stripped.viewport.map(def => def.id)).toEqual(["tablet"]);
-    expect(stripped.container).toEqual([]);
-  });
-
-  it("answers for an absent set rather than throwing", () => {
-    // The host's config states no breakpoints at all far more often than it
-    // states some, and that caller reads the field optionally.
-    expect(authoredBreakpoints(undefined)).toEqual({
-      viewport: [],
-      container: [],
-    });
-  });
-});
-
 describe("an id the compiler would drop", () => {
   it("is reported, rather than saved into nothing", () => {
     /*
@@ -368,28 +336,6 @@ describe("ids are compared as the compiler compares them", () => {
     expect(
       codes(set({ viewport: [{ id: "base", label: "B", maxWidth: 1200 }] }))
     ).toEqual(["id-reserved"]);
-  });
-});
-
-describe("cascade order", () => {
-  it("puts an unbounded definition first, then widest to narrowest", () => {
-    const ordered = inCascadeOrder([
-      { id: "narrow", label: "Narrow", maxWidth: 400 },
-      { id: "unbounded", label: "Unbounded" },
-      { id: "wide", label: "Wide", maxWidth: 900 },
-    ]);
-
-    expect(ordered.map(d => d.id)).toEqual(["unbounded", "wide", "narrow"]);
-  });
-
-  it("does not mutate the array it was given", () => {
-    const defs = [
-      { id: "a", label: "A", maxWidth: 100 },
-      { id: "b", label: "B", maxWidth: 900 },
-    ];
-    inCascadeOrder(defs);
-
-    expect(defs.map(d => d.id)).toEqual(["a", "b"]);
   });
 });
